+++ src/graphqlSchema.js
@@ -1,14 +1,18 @@
+const { AnswerModel } = require("./db/model.js");
+const postAnswer = require("./resolvers.js");
+const { logger } = require("./middlewares/logger.js")
 
+console.log(logger);
 const {
 	GraphQLID,
   GraphQLNonNull,
   GraphQLString,
   GraphQLList,
   GraphQLInt,
-
+  GraphQLSchema,
+  GraphQLObjectType,
+  GraphQLInputObjectType
 } = require("graphql");
-
-// new GraphQLList(PersonType)
 
 const answerTypeGql = {
   id: { type: GraphQLID },
@@ -51,6 +55,84 @@ const answerTypeGql = {
   travail_bureau_domicile: { type: GraphQLString },
   demo_genre: { type: GraphQLString },
   email: { type: GraphQLString },
-}; 
+};
 
- module.exports = answerTypeGql;
+const AnswerType = new GraphQLObjectType({
+	name: "Answer",
+	fields: answerTypeGql
+});
+
+const AnswerInputType = new GraphQLInputObjectType({
+	name: "AnswerInput",
+	fields: answerTypeGql
+});
+
+module.exports = new GraphQLSchema({
+	query: new GraphQLObjectType({
+		name: "Query",
+		fields: {
+			// get answer list stored id db
+			answer: {
+				type: GraphQLList(AnswerType),
+				resolve: () => {
+					return AnswerModel.find().exec()
+					.then(res => res)
+					.catch(err => {
+						logger.error(`An error occured in answer querry ${err}`);
+						return err;
+					});
+				}
+			},
+			// get an answer by id
+			answerByID: {
+				type: AnswerType,
+				args: {
+					// strong validation for graphqlid, which is mendatory for running this query
+					id: { type: GraphQLNonNull(GraphQLID) }
+				},
+				resolve: (_, args) => {
+					return AnswerModel.findById(args.id).exec()
+					.then(res => res)
+					.catch(err => {
+						logger.error(`An error occured in answerByID querry ${err}`);
+						return err;
+					});
+				}
+			},
+		}
+	}),
+	// Create Mutation
+	mutation: new GraphQLObjectType({
+		name: "Mutation",
+		fields: {
+			createAnswer: {
+				type: AnswerType,
+				args: {
+					answer: { type: (AnswerInputType) }
+				},
+				resolve: async (_, args) => {
+					return postAnswer(args["answer"])
+					.then(newAnswer => newAnswer)
+					.catch(err => {
+						logger.error(`An error occured in createAnswer mutation ${err}`);
+						return err;
+					});
+				}
+			},
+			createMultipleAnswer: {
+				type: GraphQLList(AnswerType),
+				args: {
+					answerList: { type: GraphQLList(AnswerInputType) }
+				},
+				resolve: async (_, args) => {
+					return await AnswerModel.collection.insertMany(args["answerList"])
+					.then(res => res["ops"])
+					.catch(err => {
+						logger.error(err);
+						return err;
+					});
+				}
+			}
+		}
+	})
+});
