--- src/middlewares.js
@@ -1,11 +1,0 @@
-const morgan = require('morgan');
-const logger = require('./logger.js');
-
-logger.stream = {
-  write: message => logger.info(message.substring(0, message.lastIndexOf('\n')))
-};
-
-module.exports = morgan(
-  ':method :url :status :response-time ms - :res[content-length]',
-  { stream: logger.stream }
-);
