+++ src/middlewares/logger.js
@@ -1,4 +1,5 @@
 const { createLogger, transports, format } = require('winston');
+const morgan = require('morgan');
 
 const logger = createLogger({
   format: format.combine(
@@ -16,4 +17,13 @@ const logger = createLogger({
   ]
 });
 
-module.exports = logger;
+logger.stream = {
+  write: message => logger.info(message.substring(0, message.lastIndexOf('\n')))
+};
+
+const httpLogger = morgan(
+  ':method :url :status :response-time ms - :res[content-length]',
+  { stream: logger.stream }
+);
+
+module.exports = { httpLogger, logger }
